chore(scripts): move replay to scripts
