+++ pyproject.toml
@@ -166,7 +166,7 @@ lerobot-calibrate="lerobot.calibrate:main"
 lerobot-find-cameras="lerobot.scripts.lerobot_find_cameras:main"
 lerobot-find-port="lerobot.find_port:main"
 lerobot-record="lerobot.record:main"
-lerobot-replay="lerobot.replay:main"
+lerobot-replay="lerobot.scripts.lerobot_replay:main"
 lerobot-setup-motors="lerobot.setup_motors:main"
 lerobot-teleoperate="lerobot.teleoperate:main"
 lerobot-eval="lerobot.scripts.eval:main"
